Felipe custom folder & scene - please read description
Hey guys I just added a new custom folder and a new scene called Felipe-custom. I created a very ugly room with a plane on the wall to simulate a painting. 
I created a script called PaintingTrigger that is located in my folder.
It is a simple scrip that does what we need for now. When we focus on the "painting" the name of the painting shows in the top and when we look away, it hides. Also, if we look at the painting and click, it plays a song (what I think can be some infos of the painting) and when we look away it stops. Hope it can be of some use for you guys.